flask_app/requirements.txt
Flask
Pillow
joblib
numpy
scikit-learn==1.6.1
